Did work on all my views and getting them set up with the presenters. I'm not sure I did this right though.
