+++ app/src/main/java/wonderful/workouts/dialogs/DialogHelperCallback.java
@@ -1,0 +1,7 @@
+package wonderful.workouts.dialogs;
+
+import android.content.DialogInterface;
+
+public interface DialogHelperCallback {
+    void onOkayClicked(DialogInterface dialog, int which);
+}
