+++ app/src/main/java/wonderful/workouts/dialogs/MeasurementCallback.java
@@ -1,0 +1,5 @@
+package wonderful.workouts.dialogs;
+
+public interface MeasurementCallback {
+    void onComplete(String text);
+}
