+++ app/src/main/java/wonderful/workouts/fragments/WorkoutView.java
@@ -16,16 +16,24 @@ import com.google.android.material.floatingactionbutton.FloatingActionButton;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
+import java.util.List;
 
 import wonderful.workouts.R;
 import wonderful.workouts.adapters.MovementAdapter;
+import wonderful.workouts.adapters.WorkoutAdapter;
 import wonderful.workouts.adapters.WorkoutHistoryAdapter;
 import wonderful.workouts.database.entities.Movement;
+import wonderful.workouts.database.entities.Workout;
 import wonderful.workouts.database.entities.WorkoutHistory;
+import wonderful.workouts.database.joiners.WorkoutWithHistory;
 import wonderful.workouts.databinding.FragmentWorkoutBinding;
+import wonderful.workouts.presenters.UserPresenter;
+import wonderful.workouts.presenters.WorkoutPresenter;
 
 public class WorkoutView extends Fragment {
     private FragmentWorkoutBinding binding;
+    private ListView pastWorkoutListView;
+    private ListView movementsListView;
 
     public View onCreateView(
             @NonNull LayoutInflater inflater,
@@ -36,22 +44,23 @@ public class WorkoutView extends Fragment {
         View root = binding.getRoot();
 
         //Get dummy data for now
-        ArrayList<WorkoutHistory> workoutHistories = getWorkoutHistory();
+        //ArrayList<WorkoutHistory> workoutHistories = getWorkoutHistory();
 
-        ListView workoutHistoryListView = (ListView) root.findViewById(R.id.workout_view_past_entries);
+        pastWorkoutListView = (ListView) root.findViewById(R.id.workout_view_past_entries);
+        getWorkoutHistory();
 
-        if (workoutHistoryListView == null) {
-            Log.i("WorkoutView", "History list view was null");
-        }
+//        if (workoutHistoryListView == null) {
+//            Log.i("WorkoutView", "History list view was null");
+//        }
+//
+//        workoutHistoryListView.setAdapter(new WorkoutHistoryAdapter(this.getContext(), workoutHistories));
 
-        workoutHistoryListView.setAdapter(new WorkoutHistoryAdapter(this.getContext(), workoutHistories));
+        //ArrayList<Movement> movements = getMovements();
 
-        ArrayList<Movement> movements = getMovements();
-
-        ListView movementListView = (ListView) root.findViewById(R.id.workout_history_view_movements_list);
+        movementsListView = (ListView) root.findViewById(R.id.workout_history_view_movements_list);
 
         // Set the ListView's adapter to our custom adapter!
-        movementListView.setAdapter(new MovementAdapter(this.getContext(), movements));
+        //movementListView.setAdapter(new MovementAdapter(this.getContext(), movements));
 
         //Add an event to the Floating Action Button
         FloatingActionButton btnTesting = root.findViewById(R.id.workout_new_movement);
@@ -64,7 +73,7 @@ public class WorkoutView extends Fragment {
         Button startWorkoutBtn = root.findViewById(R.id.workout_start_workout_button);
         startWorkoutBtn.setOnClickListener(view ->{
             Log.i("WorkoutView", String.format("Starting a workout"));
-            Navigation.findNavController(view).navigate(R.id.navigation_workout_active);
+            Navigation.findNavController(view).navigate(R.id.navigation_workout_history);
         });
 
         // Button btnLogin = root.findViewById(R.id.btn_login);
@@ -88,43 +97,155 @@ public class WorkoutView extends Fragment {
     }
 
     private ArrayList<WorkoutHistory> getWorkoutHistory() {
-        ArrayList<WorkoutHistory> workoutHistories = new ArrayList<>();
+        new Thread(() -> {
+            // Get the presenter.
+            UserPresenter userPresenter = UserPresenter.getInstance(requireContext());
+            WorkoutPresenter workoutPresenter = WorkoutPresenter.getInstance(requireContext());
 
-        WorkoutHistory wh = new WorkoutHistory();
-        wh.startTime = LocalDateTime.now();
-        workoutHistories.add(wh);
+            Workout workoutClicked = workoutPresenter.getCurrentWorkout();
+            WorkoutWithHistory workoutHistories = workoutPresenter.getAllPastWorkoutHistories(workoutClicked);
+            //List<Workout> workouts = workoutPresenter.getWorkoutsForUser(userPresenter.getCurrentUser());
 
-        WorkoutHistory wh1 = new WorkoutHistory();
-        wh1.startTime = LocalDateTime.now().plusDays(5);
-        workoutHistories.add(wh1);
 
-        WorkoutHistory wh2 = new WorkoutHistory();
-        wh2.startTime = LocalDateTime.now().plusDays(10);
-        workoutHistories.add(wh2);
+            // Now that we have the workouts build it on the UI thread to update the UI
+            requireActivity().runOnUiThread(() -> {
+                // Set the ListView's adapter to our custom adapter!
+                pastWorkoutListView.setAdapter(new WorkoutAdapter(this.getContext(), workoutHistories));
 
-        return workoutHistories;
+                // Add an onClick listener just for an example!
+                pastWorkoutListView.setOnItemClickListener((parent, view, position, id) -> {
+                    Workout clickedWorkout = (Workout) pastWorkoutListView.getItemAtPosition(position);
+
+                    // Store the workout in the state
+                    workoutPresenter.setCurrentWorkout(clickedWorkout);
+
+                    // new Thread(() -> {
+                    //     WorkoutHistory activeWorkout = workoutPresenter.startWorkout(clickedWorkout);
+                    //     workoutPresenter.setActiveWorkout(activeWorkout);
+                    //     requireActivity().runOnUiThread(() -> {
+                    //         Navigation.findNavController(root).navigate(R.id.navigation_workout_active);
+                    //     });
+                    // }).start();
+
+                    // Navigate to the workout page to display the workout
+                    Navigation.findNavController(root).navigate(R.id.navigation_workout);
+
+                    Log.i("HomeView", String.format("We clicked workout id: %d name: %s", clickedWorkout.workoutId, clickedWorkout.name));
+                });
+            });
+        }).start();
+
+//       new Thread(() -> {
+//            // Get the presenter.
+//            UserPresenter userPresenter = UserPresenter.getInstance(requireContext());
+//            WorkoutPresenter workoutPresenter = WorkoutPresenter.getInstance(requireContext());
+//
+//            List<Workout> workouts = workoutPresenter.getWorkoutsForUser(userPresenter.getCurrentUser());
+//
+//            // Now that we have the workouts build it on the UI thread to update the UI
+//            requireActivity().runOnUiThread(() -> {
+//                // Set the ListView's adapter to our custom adapter!
+//                workoutListView.setAdapter(new WorkoutAdapter(this.getContext(), workouts));
+//
+//                // Add an onClick listener just for an example!
+//                workoutListView.setOnItemClickListener((parent, view, position, id) -> {
+//                    Workout clickedWorkout = (Workout) workoutListView.getItemAtPosition(position);
+//
+//                    // Store the workout in the state
+//                    workoutPresenter.setCurrentWorkout(clickedWorkout);
+//
+//                    // new Thread(() -> {
+//                    //     WorkoutHistory activeWorkout = workoutPresenter.startWorkout(clickedWorkout);
+//                    //     workoutPresenter.setActiveWorkout(activeWorkout);
+//                    //     requireActivity().runOnUiThread(() -> {
+//                    //         Navigation.findNavController(root).navigate(R.id.navigation_workout_active);
+//                    //     });
+//                    // }).start();
+//
+//                    // Navigate to the workout page to display the workout
+//                    Navigation.findNavController(root).navigate(R.id.navigation_workout);
+//
+//                    Log.i("HomeView", String.format("We clicked workout id: %d name: %s", clickedWorkout.workoutId, clickedWorkout.name));
+//                });
+//            });
+//        }).start();
+//
+//
+//        ArrayList<WorkoutHistory> workoutHistories = new ArrayList<>();
+//
+//        WorkoutHistory wh = new WorkoutHistory();
+//        wh.startTime = LocalDateTime.now();
+//        workoutHistories.add(wh);
+//
+//        WorkoutHistory wh1 = new WorkoutHistory();
+//        wh1.startTime = LocalDateTime.now().plusDays(5);
+//        workoutHistories.add(wh1);
+//
+//        WorkoutHistory wh2 = new WorkoutHistory();
+//        wh2.startTime = LocalDateTime.now().plusDays(10);
+//        workoutHistories.add(wh2);
+//
+//        return workoutHistories;
     }
 
     private ArrayList<Movement> getMovements() {
-        ArrayList<Movement> movements = new ArrayList<>();
+        new Thread(() -> {
+            // Get the presenter.
+            UserPresenter userPresenter = UserPresenter.getInstance(requireContext());
+            WorkoutPresenter workoutPresenter = WorkoutPresenter.getInstance(requireContext());
 
-        Movement m = new Movement();
-        m.movementId = 1;
-        m.name = "Leg Press";
-        movements.add(m);
-
-        Movement m1 = new Movement();
-        m1.movementId = 2;
-        m1.name = "Squats";
-        movements.add(m1);
-
-        Movement m2 = new Movement();
-        m2.movementId = 3;
-        m2.name = "Calve Raises";
-        movements.add(m2);
+            Workout workoutClicked = workoutPresenter.getCurrentWorkout();
+            List<Movement> movements = workoutPresenter.getWorkoutMovements(workoutClicked);
 
 
+            // Now that we have the workouts build it on the UI thread to update the UI
+            requireActivity().runOnUiThread(() -> {
+                // Set the ListView's adapter to our custom adapter!
+                movementsListView.setAdapter(new WorkoutAdapter(this.getContext(), movements));
 
-        return movements;
+                // Add an onClick listener just for an example!
+                //movementsListView.setOnItemClickListener((parent, view, position, id) -> {
+                    //Workout clickedWorkout = (Workout) movementsListView.getItemAtPosition(position);
+
+                    // Store the workout in the state
+                    //workoutPresenter.setCurrentWorkout(clickedWorkout);
+
+                    // new Thread(() -> {
+                    //     WorkoutHistory activeWorkout = workoutPresenter.startWorkout(clickedWorkout);
+                    //     workoutPresenter.setActiveWorkout(activeWorkout);
+                    //     requireActivity().runOnUiThread(() -> {
+                    //         Navigation.findNavController(root).navigate(R.id.navigation_workout_active);
+                    //     });
+                    // }).start();
+
+                    // Navigate to the workout page to display the workout
+                    //Navigation.findNavController(root).navigate(R.id.navigation_workout);
+
+                    //Log.i("HomeView", String.format("We clicked workout id: %d name: %s", clickedWorkout.workoutId, clickedWorkout.name));
+                });
+            });
+        }).start();
+
+
+//        ArrayList<Movement> movements = new ArrayList<>();
+//
+//        Movement m = new Movement();
+//        m.movementId = 1;
+//        m.name = "Leg Press";
+//        movements.add(m);
+//
+//        Movement m1 = new Movement();
+//        m1.movementId = 2;
+//        m1.name = "Squats";
+//        movements.add(m1);
+//
+//        Movement m2 = new Movement();
+//        m2.movementId = 3;
+//        m2.name = "Calve Raises";
+//        movements.add(m2);
+//
+//
+//
+//        return movements;
     }
 }
