+++ app/src/main/java/wonderful/workouts/presenters/MotivationalQuotePresenter.java
@@ -1,0 +1,4 @@
+package wonderful.workouts.presenters;
+
+public class MotivationalQuotePresenter {
+}
